streamlit/requirements.txt
streamlit>=1.22
pandas>=1.5
plotly>=5.15
python-dotenv>=1.0
snowflake-connector-python>=3.0
cryptography>=41.0
plotly.express>=0.4.1
